requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-mistralai==0.0.1
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.24.3
